Clean up Misc.js
