+++ d2bs/kolbot/libs/bots/Duriel.js
@@ -67,7 +67,7 @@ function Duriel(Config, Attack, Pickit, Pather, Town) {
 	if (unit) {
 		for (i = 0; i < 3; i += 1) {
 			if (me.area === unit.area) {
-				Skill.cast(43, 0, unit);
+				me.cast(43, 0, unit);
 			}
 
 			if (me.area === 73) {
