+++ d2bs/kolbot/libs/bots/Travincal.js
@@ -46,7 +46,7 @@ function Travincal(Config, Attack, Pickit, Pather, Town) {
 			if (i % 4 === 0) {
 				Pather.moveTo(orgX + coords[i], orgY + coords[i + 1]);
 			} else {
-				Skill.cast(143, 0, orgX + coords[i], orgY + coords[i + 1]);
+				me.cast(143, 0, orgX + coords[i], orgY + coords[i + 1]);
 				Attack.clearList(this.buildList(1));
 			}
 		}
