+++ d2bs/kolbot/libs/common/Misc.js
@@ -4,735 +4,6 @@
 *	@desc		misc library containing Skill, Misc and Sort classes
 */
 
-var Skill = {
-	config: require('Config'),
-	usePvpRange: false,
-
-	getRange: function (skillId) {
-		switch (skillId) {
-		case 0: // Normal Attack
-			return Attack.usingBow() ? 20 : 3;
-		case 1: // Kick
-		case 5: // Left Hand Swing
-		case 10: // Jab
-		case 14: // Power Strike
-		case 19: // Impale
-		case 24: // Charged Strike
-		case 30: // Fend
-		case 34: // Lightning Strike
-		case 46: // Blaze
-		case 73: // Poison Dagger
-		case 96: // Sacrifice
-		case 97: // Smite
-		case 106: // Zeal
-		case 111: // Vengeance
-		case 112: // Blessed Hammer
-		case 116: // Conversion
-		case 126: // Bash
-		case 131: // Find Potion
-		case 133: // Double Swing
-		case 139: // Stun
-		case 142: // Find Item
-		case 144: // Concentrate
-		case 147: // Frenzy
-		case 150: // Grim Ward
-		case 152: // Berserk
-		case 232: // Feral Rage
-		case 233: // Maul
-		case 238: // Rabies
-		case 239: // Fire Claws
-		case 242: // Hunger
-		case 248: // Fury
-		case 255: // Dragon Talon
-		case 260: // Dragon Claw
-		case 270: // Dragon Tail
-			return 3;
-		case 146: // Battle Cry
-		case 154: // War Cry
-			return 4;
-		case 44: // Frost Nova
-		case 240: // Twister
-		case 245: // Tornado
-		case 500: // Summoner
-			return 5;
-		case 38: // Charged Bolt
-			return 6;
-		case 48: // Nova
-		case 151: // Whirlwind
-			return 7;
-		case 92: // Poison Nova
-			return 8;
-		case 249: // Armageddon
-			return 9;
-		case 15: // Poison Javelin
-		case 25: // Plague Javelin
-		case 101: // Holy Bolt
-		case 107: // Charge
-		case 130: // Howl
-		case 225: // Firestorm
-		case 229: // Molten Boulder
-		case 243: // Shock Wave
-			return 10;
-		case 8: // Inner Sight
-		case 17: // Slow Missiles
-			return 13;
-		case 35: // Lightning Fury
-		case 64: // Frozen Orb
-		case 67: // Teeth
-		case 234: // Fissure
-		case 244: // Volcano
-		case 251: // Fire Blast
-		case 256: // Shock Web
-		case 257: // Blade Sentinel
-		case 266: // Blade Fury
-			return 15;
-		case 7: // Fire Arrow
-		case 12: // Multiple Shot
-		case 16: // Exploding Arrow
-		case 22: // Guided Arrow
-		case 27: // Immolation Arrow
-		case 31: // Freezing Arrow
-		case 95: // Revive
-		case 121: // Fist of the Heavens
-		case 140: // Double Throw
-		case 253: // Psychic Hammer
-		case 275: // Dragon Flight
-			return 20;
-		case 91: // Lower Resist
-			return 50;
-		// Variable range
-		case 42: // Static Field
-			return Math.floor((me.getSkill(42, 1) + 4) * 2 / 3);
-		case 132: // Leap
-			var leap = [4, 7, 8, 10, 11, 12, 12, 13, 14, 14, 14, 14, 15, 16, 16, 16, 16, 16, 16, 16, 16, 16, 16, 16, 17];
-
-			return leap[Math.min(me.getSkill(132, 1) - 1, 24)];
-		case 230: // Arctic Blast
-			var arctic = [5, 6, 6, 6, 6, 7, 7, 8, 8, 8, 8, 9, 9, 10, 10, 10, 10, 11, 11, 12];
-
-			return arctic[Math.min(me.getSkill(230, 1) - 1, 19)];
-		case 49: // Lightning
-		case 84: // Bone Spear
-		case 93: // Bone Spirit
-			if (this.usePvpRange) {
-				return 40;
-			}
-
-			return 15;
-		case 47: // Fire Ball
-		case 51: // Fire Wall
-		case 53: // Chain Lightning
-		case 56: // Meteor
-		case 59: // Blizzard
-		case 273: // Mind Blast
-			if (this.usePvpRange) {
-				return 40;
-			}
-
-			return 20;
-		}
-
-		// Every other skill
-		if (this.usePvpRange) {
-			return 40;
-		}
-
-		return 20;
-	},
-
-	getHand: function (skillId) {
-		switch (skillId) {
-		case 6: // Magic Arrow
-		case 7: // Fire Arrow
-		case 9: // Critical Strike
-		case 10: // Jab
-		case 11: // Cold Arrow
-		case 12: // Multiple Shot
-		case 13: // Dodge
-		case 14: // Power Strike
-		case 15: // Poison Javelin
-		case 16: // Exploding Arrow
-		case 18: // Avoid
-		case 19: // Impale
-		case 20: // Lightning Bolt
-		case 21: // Ice Arrow
-		case 22: // Guided Arrow
-		case 23: // Penetrate
-		case 24: // Charged Strike
-		case 25: // Plague Javelin
-		case 26: // Strafe
-		case 27: // Immolation Arrow
-		case 29: // Evade
-		case 30: // Fend
-		case 31: // Freezing Arrow
-		case 33: // Pierce
-		case 34: // Lightning Strike
-		case 35: // Lightning Fury
-		case 36: // Fire Bolt
-		case 37: // Warmth
-		case 38: // Charged Bolt
-		case 39: // Ice Bolt
-		case 41: // Inferno
-		case 45: // Ice Blast
-		case 47: // Fire Ball
-		case 49: // Lightning
-		case 53: // Chain Lightning
-		case 55: // Glacial Spike
-		case 61: // Fire Mastery
-		case 63: // Lightning Mastery
-		case 64: // Frozen Orb
-		case 65: // Cold Mastery
-		case 67: // Teeth
-		case 73: // Poison Dagger
-		case 79: // Golem Mastery
-		case 84: // Bone Spear
-		case 89: // Summon Resist
-		case 93: // Bone Spirit
-		case 101: // Holy Bolt
-		case 107: // Charge
-		case 111: // Vengeance
-		case 112: // Blessed Hammer
-		case 121: // Fist of the Heavens
-		case 132: // Leap
-		case 140: // Double Throw
-		case 143: // Leap Attack
-		case 151: // Whirlwind
-		case 225: // Firestorm
-		case 229: // Molten Boulder
-		case 230: // Arctic Blast
-		case 240: // Twister
-		case 243: // Shock Wave
-		case 245: // Tornado
-		case 251: // Fire Trauma
-		case 254: // Tiger Strike
-		case 256: // Shock Field
-		case 257: // Blade Sentinel
-		case 259: // Fists of Fire
-		case 263: // Weapon Block
-		case 265: // Cobra Strike
-		case 266: // Blade Fury
-		case 269: // Claws of Thunder
-		case 274: // Blades of Ice
-		case 275: // Dragon Flight
-			return 1;
-		case 0: // Normal Attack
-		case 96: // Sacrifice
-		case 97: // Smite
-		case 106: // Zeal
-		case 116: // Conversion
-		case 126: // Bash
-		case 133: // Double Swing
-		case 139: // Stun
-		case 144: // Concentrate
-		case 147: // Frenzy
-		case 152: // Berserk
-		case 232: // Feral Rage
-		case 233: // Maul
-		case 238: // Rabies
-		case 239: // Fire Claws
-		case 242: // Hunger
-		case 248: // Fury
-		case 255: // Dragon Talon
-		case 260: // Dragon Claw
-		case 270: // Dragon Tail
-			return 2; // Shift bypass
-		}
-
-		// Every other skill
-		return 0;
-	},
-
-	charges: [],
-
-	// Cast a skill on self, Unit or coords
-	cast: function (skillId, hand, x, y, item) {
-		if (me.inTown && !this.townSkill(skillId)) {
-			return false;
-		}
-
-		if (!item && !me.getSkill(skillId, 1)) {
-			return false;
-		}
-
-		if (!this.wereFormCheck(skillId)) {
-			return false;
-		}
-
-		// Check mana cost, charged skills don't use mana
-		if (!item && this.getManaCost(skillId) > me.mp) {
-			// Maybe delay on ALL skills that we don't have enough mana for?
-			if (Skill.config.AttackSkill.concat([42, 54]).concat(Skill.config.LowManaSkill).indexOf(skillId) > -1) {
-				delay(300);
-			}
-
-			return false;
-		}
-
-		if (skillId === undefined) {
-			throw new Error("Skill.cast: Must supply a skill ID");
-		}
-
-		var i, n, clickType, shift;
-
-		if (hand === undefined) {
-			hand = 0;
-		}
-
-		if (x === undefined) {
-			x = me.x;
-		}
-
-		if (y === undefined) {
-			y = me.y;
-		}
-
-		if (!this.setSkill(skillId, hand, item)) {
-			return false;
-		}
-
-		if (Skill.config.PacketCasting > 1) {
-			switch (typeof x) {
-			case "number":
-				Packet.castSkill(hand, x, y);
-				delay(250);
-
-				break;
-			case "object":
-				Packet.unitCast(hand, x);
-				delay(250);
-
-				break;
-			}
-		} else {
-			switch (hand) {
-			case 0: // Right hand + No Shift
-				clickType = 3;
-				shift = 0;
-
-				break;
-			case 1: // Left hand + Shift
-				clickType = 0;
-				shift = 1;
-
-				break;
-			case 2: // Left hand + No Shift
-				clickType = 0;
-				shift = 0;
-
-				break;
-			case 3: // Right hand + Shift
-				clickType = 3;
-				shift = 1;
-
-				break;
-			}
-
-MainLoop:
-			for (n = 0; n < 3; n += 1) {
-				if (typeof x === "object") {
-					clickMap(clickType, shift, x);
-				} else {
-					clickMap(clickType, shift, x, y);
-				}
-
-				delay(20);
-
-				if (typeof x === "object") {
-					clickMap(clickType + 2, shift, x);
-				} else {
-					clickMap(clickType + 2, shift, x, y);
-				}
-
-				for (i = 0; i < 8; i += 1) {
-					if (me.attacking) {
-						break MainLoop;
-					}
-
-					delay(20);
-				}
-			}
-
-			while (me.attacking) {
-				delay(10);
-			}
-		}
-
-		if (this.isTimed(skillId)) { // account for lag, state 121 doesn't kick in immediately
-			for (i = 0; i < 10; i += 1) {
-				if ([4, 9].indexOf(me.mode) > -1) {
-					break;
-				}
-
-				if (me.getState(121)) {
-					break;
-				}
-
-				delay(10);
-			}
-		}
-
-		return true;
-	},
-
-	// Put a skill on desired slot
-	setSkill: function (skillId, hand, item) {
-		// Check if the skill is already set
-		if (me.getSkill(hand === 0 ? 2 : 3) === skillId) {
-			return true;
-		}
-
-		if (!item && !me.getSkill(skillId, 1)) {
-			return false;
-		}
-
-		// Charged skills must be cast from right hand
-		if (hand === undefined || hand === 3 || item) {
-			item && hand !== 0 && print('[ÿc9Warningÿc0] charged skills must be cast from right hand');
-			hand = 0;
-		}
-
-		if (me.setSkill(skillId, hand, item)) {
-			return true;
-		}
-
-		return false;
-	},
-
-	// Timed skills
-	isTimed: function (skillId) {
-		return [15, 25, 27, 51, 56, 59, 62, 64, 121, 225, 223, 228, 229, 234, 244, 247, 249, 250, 256, 268, 275, 277, 279].indexOf(skillId) > -1;
-	},
-
-	// Wereform skill check
-	wereFormCheck: function (skillId) {
-		if (!me.getState(139) && !me.getState(140)) {
-			return true;
-		}
-
-		// Can be cast by both
-		if ([0, 1, 221, 222, 226, 227, 231, 236, 237, 239, 241, 242, 246, 247, 249].indexOf(skillId) > -1) {
-			return true;
-		}
-
-		// Can be cast by werewolf only
-		if (me.getState(139) && [223, 232, 238, 248].indexOf(skillId) > -1) {
-			return true;
-		}
-
-		// Can be cast by werebear only
-		if (me.getState(140) && [228, 233, 243].indexOf(skillId) > -1) {
-			return true;
-		}
-
-		return false;
-	},
-
-	// Skills that cn be cast in town
-	townSkill: function (skillId) {
-		return [32, 40, 43, 50, 52, 58, 60, 68, 75, 85, 94, 117, 221, 222, 226, 227, 235, 236, 237, 246, 247, 258, 267, 268, 277, 278, 279].indexOf(skillId) > -1;
-	},
-
-	manaCostList: {},
-
-	// Get mana cost of the skill (mBot)
-	getManaCost: function (skillId) {
-		if (skillId < 6) {
-			return 0;
-		}
-
-		if (this.manaCostList.hasOwnProperty(skillId)) {
-			return this.manaCostList[skillId];
-		}
-
-		var skillLvl = me.getSkill(skillId, 1),
-			effectiveShift = [1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024],
-			lvlmana = getBaseStat(3, skillId, "lvlmana") === 65535 ? -1 : getBaseStat(3, skillId, "lvlmana"), // Correction for skills that need less mana with levels (kolton)
-			ret = Math.max((getBaseStat(3, skillId, "mana") + lvlmana * (skillLvl - 1)) * (effectiveShift[getBaseStat(3, skillId, "manashift")] / 256), getBaseStat(3, skillId, "minmana"));
-
-		if (!this.manaCostList.hasOwnProperty(skillId)) {
-			this.manaCostList[skillId] = ret;
-		}
-
-		return ret;
-	}
-};
-
-var Item = {
-	hasTier: function (item) {
-		const NTIP = require('NTIP');
-		return Misc.config.AutoEquip && NTIP.GetTier(item) > 0;
-	},
-
-	canEquip: function (item) {
-		if (item.type !== 4) { // Not an item
-			return false;
-		}
-
-		if (!item.getFlag(0x10)) { // Unid item
-			return false;
-		}
-
-		if (item.getStat(92) > me.getStat(12) || item.dexreq > me.getStat(2) || item.strreq > me.getStat(0)) { // Higher requirements
-			return false;
-		}
-
-		return true;
-	},
-
-	// Equips an item and throws away the old equipped item
-	equip: function (item, bodyLoc) {
-		if (!this.canEquip(item)) {
-			return false;
-		}
-		const Town = require('Town');
-
-		// Already equipped in the right slot
-		if (item.mode === 1 && item.bodylocation === bodyLoc) {
-			return true;
-		}
-
-		var i, cursorItem;
-
-		if (item.location === 7) {
-			if (!Town.openStash()) {
-				return false;
-			}
-		}
-
-		const Storage = require('Storage');
-		for (i = 0; i < 3; i += 1) {
-			if (item.toCursor()) {
-				clickItem(0, bodyLoc);
-				delay(me.ping * 2 + 500);
-
-				if (item.bodylocation === bodyLoc) {
-					if (getCursorType() === 3) {
-						//Misc.click(0, 0, me);
-
-						cursorItem = getUnit(100);
-
-						if (cursorItem) {
-							if (!Storage.Inventory.CanFit(cursorItem) || !Storage.Inventory.MoveTo(cursorItem)) {
-								cursorItem.drop();
-							}
-						}
-					}
-
-					return true;
-				}
-			}
-		}
-
-		return false;
-	},
-
-	getEquippedItem: function (bodyLoc) {
-		var item = me.getItem();
-		const NTIP = require('NTIP');
-		if (item) {
-			do {
-				if (item.bodylocation === bodyLoc) {
-					return {
-						classid: item.classid,
-						tier: NTIP.GetTier(item)
-					};
-				}
-			} while (item.getNext());
-		}
-
-		// Don't have anything equipped in there
-		return {
-			classid: -1,
-			tier: -1
-		};
-	},
-
-	getBodyLoc: function (item) {
-		var bodyLoc;
-
-		switch (item.itemType) {
-		case 2: // Shield
-		case 70: // Auric Shields
-			bodyLoc = 5;
-
-			break;
-		case 3: // Armor
-			bodyLoc = 3;
-
-			break;
-		case 5: // Arrows
-		case 6: // Bolts
-			bodyLoc = 5;
-
-			break;
-		case 10: // Ring
-			bodyLoc = [6, 7];
-
-			break;
-		case 12: // Amulet
-			bodyLoc = 2;
-
-			break;
-		case 15: // Boots
-			bodyLoc = 9;
-
-			break;
-		case 16: // Gloves
-			bodyLoc = 10;
-
-			break;
-		case 19: // Belt
-			bodyLoc = 8;
-
-			break;
-		case 37: // Helm
-		case 71: // Barb Helm
-		case 75: // Circlet
-			bodyLoc = 1;
-
-			break;
-		case 24: //
-		case 25: //
-		case 26: //
-		case 27: //
-		case 28: //
-		case 29: //
-		case 30: //
-		case 31: //
-		case 32: //
-		case 33: //
-		case 34: //
-		case 35: //
-		case 36: //
-		case 42: //
-		case 43: //
-		case 44: //
-		case 67: // Handtohand (Assasin Claw)
-		case 68: //
-		case 69: //
-		case 72: //
-		case 85: //
-		case 86: //
-		case 87: //
-		case 88: //
-			bodyLoc = 4;
-
-			break;
-		default:
-			return false;
-		}
-
-		if (typeof bodyLoc === "number") {
-			bodyLoc = [bodyLoc];
-		}
-
-		return bodyLoc;
-	},
-
-	autoEquipCheck: function (item) {
-		if (!Misc.config.AutoEquip) {
-			return true;
-		}
-		const NTIP = require('NTIP');
-		var i,
-			tier = NTIP.GetTier(item),
-			bodyLoc = this.getBodyLoc(item);
-
-		if (tier > 0 && bodyLoc) {
-			for (i = 0; i < bodyLoc.length; i += 1) {
-				// Low tier items shouldn't be kept if they can't be equipped
-				if (tier > this.getEquippedItem(bodyLoc[i]).tier && (this.canEquip(item) || !item.getFlag(0x10))) {
-					return true;
-				}
-			}
-		}
-
-		// Sell/ignore low tier items, keep high tier
-		if (tier > 0 && tier < 100) {
-			return false;
-		}
-
-		return true;
-	},
-
-	// returns true if the item should be kept+logged, false if not
-	autoEquip: function () {
-		if (!Misc.config.AutoEquip) {
-			return true;
-		}
-		const NTIP = require('NTIP');
-
-		var i, j, tier, bodyLoc, tome, gid,
-			items = me.findItems(-1, 0);
-
-		if (!items) {
-			return false;
-		}
-
-		function sortEq(a, b) {
-			if (Item.canEquip(a)) {
-				return -1;
-			}
-
-			if (Item.canEquip(b)) {
-				return 1;
-			}
-
-			return 0;
-		}
-
-		me.cancel();
-
-		// Remove items without tier
-		for (i = 0; i < items.length; i += 1) {
-			if (NTIP.GetTier(items[i]) === 0) {
-				items.splice(i, 1);
-
-				i -= 1;
-			}
-		}
-		const Town = require('Town');
-		while (items.length > 0) {
-			items.sort(sortEq);
-
-			tier = NTIP.GetTier(items[0]);
-			bodyLoc = this.getBodyLoc(items[0]);
-
-			if (tier > 0 && bodyLoc) {
-				for (j = 0; j < bodyLoc.length; j += 1) {
-					if ([3, 7].indexOf(items[0].location) > -1 && tier > this.getEquippedItem(bodyLoc[j]).tier && this.getEquippedItem(bodyLoc[j]).classid !== 174) { // khalim's will adjustment
-						if (!items[0].getFlag(0x10)) { // unid
-							tome = me.findItem(519, 0, 3);
-
-							if (tome && tome.getStat(70) > 0) {
-								if (items[0].location === 7) {
-									Town.openStash();
-								}
-
-								Town.identifyItem(items[0], tome);
-							}
-						}
-
-						gid = items[0].gid;
-
-						print(items[0].name);
-
-						if (this.equip(items[0], bodyLoc[j])) {
-							Misc.logItem("Equipped", me.getItem(-1, -1, gid));
-						}
-
-						break;
-					}
-				}
-			}
-
-			items.shift();
-		}
-
-		return true;
-	}
-};
-
 var Misc = {
 	config: require('Config'),
 	// Click something
@@ -1910,323 +1181,4 @@ MainLoop:
 
 		return flags.length ? flags : null;
 	}
-};
-
-var Packet = {
-	openMenu: function (unit) {
-		if (unit.type !== 1) {
-			throw new Error("openMenu: Must be used on NPCs.");
-		}
-
-		if (getUIFlag(0x08)) {
-			return true;
-		}
-
-		var i, tick;
-
-		for (i = 0; i < 5; i += 1) {
-			if (getDistance(me, unit) > 4) {
-				Pather.moveToUnit(unit);
-			}
-
-			sendPacket(1, 0x13, 4, 1, 4, unit.gid);
-			tick = getTickCount();
-
-			while (getTickCount() - tick < 5000) {
-				if (getUIFlag(0x08)) {
-					delay(Math.max(500, me.ping * 2));
-
-					return true;
-				}
-
-				if (getInteractedNPC() && getTickCount() - tick > 1000) {
-					me.cancel();
-				}
-
-				delay(100);
-			}
-
-			sendPacket(1, 0x2f, 4, 1, 4, unit.gid);
-			delay(me.ping * 2);
-			sendPacket(1, 0x30, 4, 1, 4, unit.gid);
-			delay(me.ping * 2);
-			this.flash(me.gid);
-		}
-
-		return false;
-	},
-
-	startTrade: function (unit, mode) {
-		if (unit.type !== 1) {
-			throw new Error("Unit.startTrade: Must be used on NPCs.");
-		}
-
-		if (getUIFlag(0x0C)) {
-			return true;
-		}
-
-		var i,
-			gamble = mode === "Gamble";
-
-		if (this.openMenu(unit)) {
-			for (i = 0; i < 10; i += 1) {
-				delay(200);
-
-				if (i % 2 === 0) {
-					sendPacket(1, 0x38, 4, gamble ? 2 : 1, 4, unit.gid, 4, 0);
-				}
-
-				if (unit.itemcount > 0) {
-					delay(200);
-
-					return true;
-				}
-			}
-		}
-
-		return false;
-	},
-
-	buyItem: function (unit, shiftBuy, gamble) {
-		var i, tick,
-			oldGold = me.getStat(14) + me.getStat(15),
-			itemCount = me.itemcount,
-			npc = getInteractedNPC();
-
-		if (!npc) {
-			throw new Error("buyItem: No NPC menu open.");
-		}
-
-		if (me.getStat(14) + me.getStat(15) < unit.getItemCost(0)) { // Can we afford the item?
-			return false;
-		}
-
-		for (i = 0; i < 3; i += 1) {
-			sendPacket(1, 0x32, 4, npc.gid, 4, unit.gid, 4, shiftBuy ? 0x80000000 : gamble ? 0x2 : 0x0, 4, 0);
-
-			tick = getTickCount();
-
-			while (getTickCount() - tick < Math.max(2000, me.ping * 2 + 500)) {
-				if (shiftBuy && me.getStat(14) + me.getStat(15) < oldGold) {
-					return true;
-				}
-
-				if (itemCount !== me.itemcount) {
-					return true;
-				}
-
-				delay(10);
-			}
-		}
-
-		return false;
-	},
-
-	sellItem: function (unit) {
-		if (unit.type !== 4) { // Check if it's an item we want to buy
-			throw new Error("Unit.sell: Must be used on items.");
-		}
-
-		var i, tick, npc,
-			itemCount = me.itemcount;
-
-		npc = getInteractedNPC();
-
-		if (!npc) {
-			return false;
-		}
-
-		for (i = 0; i < 5; i += 1) {
-			sendPacket(1, 0x33, 4, npc.gid, 4, unit.gid, 4, 0, 4, 0);
-
-			tick = getTickCount();
-
-			while (getTickCount() - tick < 2000) {
-				if (me.itemcount !== itemCount) {
-					return true;
-				}
-
-				delay(10);
-			}
-		}
-
-		return false;
-	},
-
-	identifyItem: function (unit, tome) {
-		var i, tick;
-
-		if (!unit || unit.getFlag(0x10)) {
-			return false;
-		}
-
-CursorLoop:
-		for (i = 0; i < 3; i += 1) {
-			sendPacket(1, 0x27, 4, unit.gid, 4, tome.gid);
-
-			tick = getTickCount();
-
-			while (getTickCount() - tick < 2000) {
-				if (getCursorType() === 6) {
-					break CursorLoop;
-				}
-
-				delay(10);
-			}
-		}
-
-		if (getCursorType() !== 6) {
-			return false;
-		}
-
-		for (i = 0; i < 3; i += 1) {
-			if (getCursorType() === 6) {
-				sendPacket(1, 0x27, 4, unit.gid, 4, tome.gid);
-			}
-
-			tick = getTickCount();
-
-			while (getTickCount() - tick < 2000) {
-				if (unit.getFlag(0x10)) {
-					delay(50);
-
-					return true;
-				}
-
-				delay(10);
-			}
-		}
-
-		return false;
-	},
-
-	itemToCursor: function (item) {
-		var i, tick;
-
-		if (me.itemoncursor) { // Something already on cursor
-			if (getUnit(100).gid === item.gid) { // Return true if the item is already on cursor
-				return true;
-			}
-
-			this.dropItem(getUnit(100)); // If another item is on cursor, drop it
-		}
-
-		for (i = 0; i < 15; i += 1) {
-			if (item.mode === 1) { // equipped
-				sendPacket(1, 0x1c, 2, item.bodylocation);
-			} else {
-				sendPacket(1, 0x19, 4, item.gid);
-			}
-
-			tick = getTickCount();
-
-			while (getTickCount() - tick < Math.max(500, me.ping * 2 + 200)) {
-				if (me.itemoncursor) {
-					return true;
-				}
-
-				delay(10);
-			}
-		}
-
-		return false;
-	},
-
-	dropItem: function (item) {
-		var i, tick;
-
-		if (!this.itemToCursor(item)) {
-			return false;
-		}
-
-		for (i = 0; i < 15; i += 1) {
-			sendPacket(1, 0x17, 4, item.gid);
-
-			tick = getTickCount();
-
-			while (getTickCount() - tick < Math.max(500, me.ping * 2 + 200)) {
-				if (!me.itemoncursor) {
-					return true;
-				}
-
-				delay(10);
-			}
-		}
-
-		return false;
-	},
-
-	castSkill: function (hand, wX, wY) {
-		hand = (hand === 0) ? 0x0c : 0x05;
-		sendPacket(1, hand, 2, wX, 2, wY);
-	},
-
-	unitCast: function (hand, who) {
-		hand = (hand === 0) ? 0x11 : 0x0a;
-		sendPacket(1, hand, 4, who.type, 4, who.gid);
-	},
-
-	moveNPC: function (npc, dwX, dwY) { // commented the patched packet
-		//sendPacket(1, 0x59, 4, npc.type, 4, npc.gid, 4, dwX, 4, dwY);
-	},
-
-	teleWalk: function (x, y, maxDist) {
-		var i;
-
-		if (maxDist === undefined) {
-			maxDist = 5;
-		}
-
-		if (!this.telewalkTick) {
-			this.telewalkTick = 0;
-		}
-
-		if (getDistance(me, x, y) > 10 && getTickCount() - this.telewalkTick > 3000 && Attack.validSpot(x, y)) {
-			for (i = 0; i < 5; i += 1) {
-				sendPacket(1, 0x5f, 2, x + rand(-1, 1), 2, y + rand(-1, 1));
-				delay(me.ping + 1);
-				sendPacket(1, 0x4b, 4, me.type, 4, me.gid);
-				delay(me.ping + 1);
-
-				if (getDistance(me, x, y) < maxDist) {
-					delay(200);
-
-					return true;
-				}
-			}
-
-			this.telewalkTick = getTickCount();
-		}
-
-		return false;
-	},
-
-	flash: function (gid, wait = 300 + 2 * me.ping) {
-		sendPacket(1, 0x4b, 4, 0, 4, gid);
-
-		if (wait > 0) {
-			delay(wait);
-		}
-	},
-
-	changeStat: function (stat, value) {
-		if (value > 0) {
-			getPacket(1, 0x1d, 1, stat, 1, value);
-		}
-	},
-
-	addListener: function (packetType, callback) { // specialized wrapper for addEventListener
-		if (typeof packetType === 'number')
-			packetType = [packetType];
-
-		if (typeof packetType === 'object' && packetType.length) {
-			addEventListener('gamepacket', packet => (packetType.indexOf(packet[0]) > -1 ? callback(packet) : false));
-
-			return callback;
-		}
-
-		return null;
-	},
-
-	removeListener: callback => removeEventListener('gamepacket', callback), // just a wrapper
 };
